Add individual collection pages and ability to view collections of public users - also some component dividing
