+++ server/models/record.js
@@ -4,7 +4,7 @@ const filePluginLib = require('mongoose-file');
 const filePlugin = filePluginLib.filePlugin;
 
 const schema = new Schema({
-  user: {
+  userId: {
     type: Schema.Types.ObjectId,
     required: false, // Should required when register/signin is implemented
     minlength: 1,
